Renamed l to listen
