--- src/l/index.js
@@ -1,11 +1,0 @@
-import Listener from './Listener';
-
-/**
- * @param {Component} component
- * @param {...string} path
- * @returns {Listener}
- */
-const l = (component, ...path) =>
-  new Listener(component, ...path);
-
-export default l;
