+++ src/r/setStyles.js
@@ -1,4 +1,4 @@
-import Listener from '../l/Listener';
+import Listener from '../listen/Listener';
 import AttributeListener from './utils/AttributeListener';
 
 /**
